[Chore] Fix console err with myBagPage
split presenter logic, use material icon for gear, fix error MUI
