+++ frontend/src/scss/components/_sortbar.scss
@@ -1,0 +1,25 @@
+.sort-bar-container {
+  position: relative;
+  z-index: 2;
+}
+
+.tab-price-container {
+  position: absolute;
+  right: 2%;
+  bottom: 20%;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-end;
+  text-align: right;
+  z-index: 1;
+
+  .tab-description {
+    @include font-default(400, 0.8rem);
+    color: gray;
+  }
+
+  .tab-price {
+    @include font-default(400, 1rem);
+    color: #f5626c;
+  }
+}
